+++ packages/framework/src/Component/DataFixture/AbstractNativeFixture.php
@@ -3,32 +3,24 @@
 namespace Shopsys\FrameworkBundle\Component\DataFixture;
 
 use Doctrine\Common\DataFixtures\AbstractFixture;
+use Doctrine\ORM\EntityManager;
 use Doctrine\ORM\Query\ResultSetMapping;
-use Symfony\Component\DependencyInjection\ContainerAwareInterface;
-use Symfony\Component\DependencyInjection\ContainerInterface;
 
-abstract class AbstractNativeFixture extends AbstractFixture implements ContainerAwareInterface
+abstract class AbstractNativeFixture extends AbstractFixture
 {
-    /**
-     * @var \Symfony\Component\DependencyInjection\ContainerInterface
-     */
-    private $container;
 
     /**
-     * @param \Symfony\Component\DependencyInjection\ContainerInterface $container
+     * @var \Doctrine\ORM\EntityManager
      */
-    public function setContainer(ContainerInterface $container = null)
-    {
-        $this->container = $container;
-    }
+    private $entityManager;
 
     /**
-     * @param string $serviceId
-     * @return mixed
+     * @required
+     * @param \Doctrine\ORM\EntityManager $entityManager
      */
-    protected function get($serviceId)
+    public function autowireEntityManager(EntityManager $entityManager)
     {
-        return $this->container->get($serviceId);
+        $this->entityManager = $entityManager;
     }
 
     /**
@@ -38,10 +30,7 @@ abstract class AbstractNativeFixture extends AbstractFixture implements Containe
      */
     protected function executeNativeQuery($sql, array $parameters = null)
     {
-        $em = $this->container->get('doctrine.orm.entity_manager');
-        /* @var $em \Doctrine\ORM\EntityManager */
-
-        $nativeQuery = $em->createNativeQuery($sql, new ResultSetMapping());
+        $nativeQuery = $this->entityManager->createNativeQuery($sql, new ResultSetMapping());
         return $nativeQuery->execute($parameters);
     }
 }