+++ packages/framework/src/DataFixtures/Base/DomainDbFunctionsDataFixture.php
@@ -8,13 +8,22 @@ use Shopsys\FrameworkBundle\Component\Domain\DomainDbFunctionsFacade;
 
 class DomainDbFunctionsDataFixture extends AbstractNativeFixture
 {
+    /** @var \Shopsys\FrameworkBundle\Component\Domain\DomainDbFunctionsFacade */
+    private $domainDbFunctionsFacade;
+
+    /**
+     * @param \Shopsys\FrameworkBundle\Component\Domain\DomainDbFunctionsFacade $domainDbFunctionsFacade
+     */
+    public function __construct(DomainDbFunctionsFacade $domainDbFunctionsFacade)
+    {
+        $this->domainDbFunctionsFacade = $domainDbFunctionsFacade;
+    }
+
     /**
      * @param \Doctrine\Common\Persistence\ObjectManager $manager
      */
     public function load(ObjectManager $manager)
     {
-        $domainDbFunctionsFacade = $this->get(DomainDbFunctionsFacade::class);
-        /* @var $domainDbFunctionsFacade \Shopsys\FrameworkBundle\Component\Domain\DomainDbFunctionsFacade */
-        $domainDbFunctionsFacade->createDomainDbFunctions();
+        $this->domainDbFunctionsFacade->createDomainDbFunctions();
     }
 }