+++ packages/framework/src/DataFixtures/Demo/TopProductDataFixture.php
@@ -11,6 +11,19 @@ use Shopsys\FrameworkBundle\Model\Product\TopProduct\TopProductFacade;
 class TopProductDataFixture extends AbstractReferenceFixture implements DependentFixtureInterface
 {
     /**
+     * @var \Shopsys\FrameworkBundle\Model\Product\TopProduct\TopProductFacade
+     */
+    private $topProductFacade;
+
+    /**
+     * @param \Shopsys\FrameworkBundle\Model\Product\TopProduct\TopProductFacade $topProductFacade
+     */
+    public function __construct(TopProductFacade $topProductFacade)
+    {
+        $this->topProductFacade = $topProductFacade;
+    }
+
+    /**
      * @param \Doctrine\Common\Persistence\ObjectManager $manager
      */
     public function load(ObjectManager $manager)
@@ -29,15 +42,12 @@ class TopProductDataFixture extends AbstractReferenceFixture implements Dependen
      */
     private function createTopProducts(array $productReferenceNames)
     {
-        $topProductFacade = $this->get(TopProductFacade::class);
-        /* @var $topProductFacade \Shopsys\FrameworkBundle\Model\Product\TopProduct\TopProductFacade */
-
         $products = [];
         foreach ($productReferenceNames as $productReferenceName) {
             $products[] = $this->getReference($productReferenceName);
         }
 
-        $topProductFacade->saveTopProductsForDomain(Domain::FIRST_DOMAIN_ID, $products);
+        $this->topProductFacade->saveTopProductsForDomain(Domain::FIRST_DOMAIN_ID, $products);
     }
 
     /**