+++ packages/framework/src/DataFixtures/Demo/UnitDataFixture.php
@@ -11,6 +11,17 @@ class UnitDataFixture extends AbstractReferenceFixture
 {
     const UNIT_CUBIC_METERS = 'unit_m3';
 
+    /** @var \Shopsys\FrameworkBundle\Model\Product\Unit\UnitFacade */
+    private $unitFacade;
+
+    /**
+     * @param \Shopsys\FrameworkBundle\Model\Product\Unit\UnitFacade $unitFacade
+     */
+    public function __construct(UnitFacade $unitFacade)
+    {
+        $this->unitFacade = $unitFacade;
+    }
+
     /**
      * @param \Doctrine\Common\Persistence\ObjectManager $manager
      */
@@ -28,10 +39,7 @@ class UnitDataFixture extends AbstractReferenceFixture
      */
     private function createUnit(UnitData $unitData, $referenceName = null)
     {
-        $unitFacade = $this->get(UnitFacade::class);
-        /* @var $unitFacade \Shopsys\FrameworkBundle\Model\Product\Unit\UnitFacade */
-
-        $unit = $unitFacade->create($unitData);
+        $unit = $this->unitFacade->create($unitData);
         if ($referenceName !== null) {
             $this->addReference($referenceName, $unit);
         }